+++ testing/e2e/skr/provisioning-service-test/internal/model.go
